+++ NEWS.md
@@ -1,3 +1,11 @@
+# strata 1.4.3
+
+## Bug Fixes
+
+* `adhoc()` will now properly match strata (#58)
+
+* Script execution ordering will now behave when there are many scripts (#59).
+
 # strata 1.4.0
 
 This release is all about bug fixes and improvements to the user experience.